requirements.txt
# Pinned, "reasonable" versions as of 2026-02; keep torch/torchaudio aligned.
#
# Note: Installing torch/torchaudio with GPU support may require following
# PyTorch's official instructions for your CUDA version.
numpy==1.26.4
PyYAML==6.0.1
requests==2.32.3
tqdm==4.66.4
soundfile==0.12.1

torch==2.2.2
torchaudio==2.2.2

